+++ src/main/java/com/workable_sb/workable/repositories/BeneficioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.Beneficio;
 
-public interface BeneficioRepository extends JpaRepository<Beneficio, Short>{
+public interface BeneficioRepository extends JpaRepository<Beneficio, Integer>{
 
 }
